fix: Address PR review — extract sandbox tests, add guard comment, update architecture docs [Midtown !2022]
